[selinux] Use semanage and only run if selinux is not disabled
Updates the selinux plugin to only run the builk of selinux commands if
selinux is not in a disabled state, as otherwise these commands do not
produce useful output. The output of sestatus is still collected even if
selinux is disabled so analysts still have a reference point for that
information.

Additionally, change to using semanage to collect the bulk of the data
given that that is the recommended practice. Collection of semanage list
output is now always collected, instead of being enabled by the 'list'
option.

Resolves: #1209

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>
Signed-off-by: Bryn M. Reeves <bmr@redhat.com>