[ArchLinux] Create new ArchPlugin tagging class.
The Arch policy sets this new tagging class as its only valid subclass
and the Pacman plugin now subclasses this new tagging class. With this
commit the Pacman plugin is the only plugin that gets executed on an
Arch system. Support for additional plugins will be added in subsequent
commits, after testing.

Additonal minor changes include the removal of an unused import from
the Arch policy and formatting changes in the Pacman plugin.

Signed-off-by: Julian Phillips <phillips.julian@gmail.com>