[policy] add optional 'release' value for packages
An a new optional field to the data gathered for installed packages
to include the 'release' string provided by the platform package
manager, and implement a change to the redhat policy family to
add this to the RPM query command for those distributions.

Policies wishing to implement this just need to ensure that there
is an additional '|' separated field in the data returned by the
specified query command. If the additional field is present it is
assumed to contain a release value and is stored in the package
dictionary for the 'release' key.

For policies that do not define the additional field in query
commands the 'release' key will always be set to None.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>