[archive] Change log level for missing compression tools
Changes the log level from error to info when sos is checking for valid
compression tools (xz, bzip2, gzip) and finds that one does not exist.

The fact that one of the compression methods isn't available is not
necessarily an error depending on where sos is being run, for example in
a container that provides xz but not gzip. In this case an error is
printed saying gzip is not available, even if sos would never have used
gzip due to the presence of xz, which could cause end user confusion.

If no supported compression methods are available, then that is still
properly reported as an error.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>