Tagged 22 more plugins with the ArchPlugin class.
These newly tagged plugins execute as is and produce useful output on
an Arch Linux system without changes to plugin execution code. The
ArchPlugin tagging class is simply imported and appended to the
existing list of derived classes in each plugin. This brings the total
number of classes tagged with the ArchPlugin to 60.

Signed-off-by: Julian Phillips <phillips.julian@gmail.com>