[sosreport] adds a per-policy package manager file list
Add a 'files' member to the PackageManager class, and a new
'files_command' kwarg to the initialiser. If the 'files_command'
is defined it is used to obtain a list of all package managed
files present on the system.

The list may be retrieved by calling PackageManager.all_files().

An implementation of 'files_command' is added to the Red Hat
policy classes.

Signed-off-by: Pierguido Lambri <plambri@redhat.com>
Signed-off-by: Bryn M. Reeves <bmr@redhat.com>