[unpackaged] add plugin
This plugin will collect a list of all binary files that are
available in the standard $PATH and that are not managed by the
distribution package manager.

For now this works only for the distributions that use the
RedHatPolicy (RHEL and Fedora): other policies may add support by
implementing the 'files_command' policy initialiser callback, and
adding the relevant tagging class to the plugin.

Resolves: #1066

Signed-off-by: Pierguido Lambri <plambri@redhat.com>
Signed-off-by: Bryn M. Reeves <bmr@redhat.com>