[openstack_instack] add support for containerized undercloud
1) Do the copy in one call with variables, so it's more efficient.

2) Collect the config tarball
It'll be very useful to collect the config tarball which containers
TripleO Heat Templates used to deploy the undercloud and the Ansible
playbooks generated by config-download as well.
Both of them are archived and compressed in
/home/stack/undercloud-install-*.tar.bzip2 files everytime
install/upgrade commands are run.

3) Collect ephemeral Heat logs
/var/log/heat-launcher is a new directory that contains the
heat-installer sqlite, logs and configuration.
The content is small because it's only used once and it's ephemeral
(purged everytime an operator install or upgrade an undercloud).
However, collecting these files will be useful to help customers.

4) Collect new logfile
When containerizing the undercloud, we're moving install-undercloud.log
into $HOME which makes it more accessible for the users, while we're
removing instack(-undercloud) bits over time.
This patch add $HOME/install-undercloud.log file and document why we're
doing this. Note that we still collect old file for previous releases.

Signed-off-by: Emilien Macchi <emilien@redhat.com>