[ovn] Add new plugins for Open Virtual Network
OVN is a sub-project of Openvswitch used to define logical networks for
a cluster of OVS instances. The two plugins defined here are
"ovn-central", which runs on a single server, and "ovn-host" which runs
on each of the hypervisors running OVS.

These plugins gather runtime information about the configured virtual
networks.

Resolves: #1259

Signed-off-by: Mark Michelson <mmichels@redhat.com>
Signed-off-by: Bryn M. Reeves <bmr@redhat.com>