[logs] Revert "[logs] implement --all-logs for Red Hat messages* and secure*"
commit 5d25dd31e731cc0d2d09a1b242803373937b4863 ("[logs] implement
--all-logs for Red Hat messages* and secure*") dropped rotated
/var/log/messages files by default.  This will make support enginners
ask their customers to collect them separately again and again because
these are the most important and the most freqently referenced files
for support engineers.  So it should be limited only with sizelimit
by default.

Also, if --all-logs is specified and the variable "messages" ends with
"*", os.path.exists(messages) always returns False, and it collects
journal unexpectedly.  So "*" should not be added to the variable
"messages" itself.

Signed-off-by: Kazuhito Hagio <k-hagio@ab.jp.nec.com>