[sosreport] Add mechanism to encrypt final archive
Adds an option to encrypt the resulting archive that sos generates.
There are two methods for doing so:

	--encrypt-key	Uses a key-pair for asymmetric encryption
	--encrypt-pass  Uses a password for symmetric encryption

For key-pair encryption, the key-to-be-used must be imported into the
root user's keyring, as gpg does not allow for the use of keyfiles.

If the encryption process fails, sos will not abort as the unencrypted
archive will have already been created. The assumption being that the
archive is still of use and/or the user has another means of encrypting
it.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>