[sos] add automatic preset detection infrastructure
Add a new special value for the --preset argument, 'auto', which
enables automatic detection of the preset identifier string by
the policy class.

A new method, probe_preset() is added which should be implemented
by derived classes of Policy to return a ProductDefaults() object
initialised appropriately for the running preset.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>