[sosreport] remove deque() use from sos
The reasoning for the use of deques in sos.sosreport is lost to
the mists of CVS, but may have had something to do with limitations
in setting default values in old versions of optparse: nothing in
the module appears to depend on deque-like behaviour at all.

Pull it all out and replace them with plain lists. This simplifies
JSON handling and option merging for these types.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>