[sosreport] use short option -v for verbosity in --list-presets
The --verbosity arg is a bit awkward in preset listings as it is
a count type. The literal value is 1, 2, or 3, but the user enters
it by repeating either the short or long option.

Ideally, a user should be able to paste the arguments from the
--list-presets output into the terminal and use them as arguments
to sosreport: using the long option for this is... verbose:
"--verbosity --verbosity --verbosity". Special case this option
and use the short version instead: "-vvv".

         name: vvverbose-debug
  description:
      options: --debug -vvv

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>