[filesys] Add permission and selinux check for /tmp
Adds a check for the /tmp directory for permissions and selinux
context. If either of these are incorrect, the system in question can
exhibit strange behavior and otherwise be difficult for support teams to
detect.

On systems where SELinux is disabled, the permission check is still
relevant and the command still produces useful output with the -Z
option.

Resolves: #1327

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>
Signed-off-by: Bryn M. Reeves <bmr@redhat.com>