[docker] Simplify command building
An early change to the docker plugin corrected the Debian/Ubuntu docker
command to be "docker" instead of "docker.io". This means that the
docker commands run by the plugin are not going to be different across
distributions. As such, the way the commands are built by the plugin
have been simplified to ease readability.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>