[sosreport] Fix threads execution with no options and output formatting
Fixes an issue where sosreport would not run due to an unset number of
threads if 'sosreport' was run bare, with no options. Any option
(threads or otherwise) would mask this problem due to how the default
threads value was set by the parser. Now sosreport can be run without
any options and run normally, with the concurrency provided by the
threads.

Additionally, changes the status line printed during plugin execution to
be of a fixed length, so that output lines are not mangled when the
status line is updated, rather than a new line being printed (such as
what happens when using --batch).

Resolves: #1336

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>
Signed-off-by: Bryn M. Reeves <bmr@redhat.com>