[sosreport] Correct verbosity level when -v is not specified
Fixes an issue where verbosity would be set to 'None' rather than 0 in
the event that a sosreport was run with options, when those options did
not include -v at all. This meant that if any options were specified, a
newline would be printed for each started plugin when instead we should
still have been refreshing a single line.

Related to this, the previous newline behavior was beneficial when
running with --batch, so that behavior is maintained if --batch is
specified.

Resolves: #1355

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>
Signed-off-by: Bryn M. Reeves <bmr@redhat.com>