[sosreport] handle OSError in XmlReport.report()
The XmlReport class in sos.sosreport traverses the list of plugins
and prepares a report including each collected file.

The access to the file is wrapped in an except block but the recent
change to pycodestyle modified this from a 'naked' except into a
'except IOError' block.

Attempting to open a non-existent path yields an IOError, however
attempting to stat a non-existent path (or a dangling symlink
without lstat()) yields OSError:

  # ln -s foo qux
  # cat qux
  cat: qux: No such file or directory

  >>> os.stat("/foo/qux")
  Traceback (most recent call last):
    File "<stdin>", line 1, in <module>
  OSError: [Errno 2] No such file or directory: '/foo/qux'

Catch both exception classes in this method.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>