[logs] restore secure* and messages* log collection
Commit 5d25dd3 "fixed" a longstanding problem with the RedHatLogs
plugin, in that the normal --all-logs behaviour (only collect the
current log unless the option is given) is not respected for the
messages and secure files.

Based on feedback from contributors and support engineers this
has now been reversed and all copies of these logs will now be
included in reports by default again.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>