[logs] fix RedHatLogs syslog/journal handling
Fix the detection of syslog logs in RedHatLogs.setup() by storing
the absolute path ("/var/log/messages"), and using this to detect
the presence of the path with os.path.exists() and append the "*"
in-place when making the add_copy_spec() call.

Also improve the readability of the method and use an explicit
variable, "have_messages", to indicate pressence of syslog style
logs.

Resolves: #1312

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>