[archive] fix add_string()/do_*_sub() regression
A change in the handling of add_string() operations in the archive
class causes the Plugin string substitution methods to fail (since
the archive was enforcing a check that the path did not already
exist - for substitutions this is always the case).

Maintain the check for content that is being copied into the
archive anew, but make the add_string() method override this and
disable the existence checks.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>