[multipath] add multipath -t output
The 'multipath -t' command prints the internal hardware table
of device-mapper-multipath. These are the true, hard-coded
defaults that multipath will follow if no other configuration
is set.  The hardware table is specific to the
device-mapper-multipath version that is in use, so it is subject
to change across release, making it difficult to track.  Prior to
the -t flag creation, the user would have to check the source
package or /usr/share/doc/device-mapper-multipath for the table.

The output is needed as a piece of the puzzle in how the
'multipathd show config' command presents the current, running
configuration (it uses the .conf and the hardware table). This
makes it very useful in troubleshooting activities.

Resolves: #1379

Signed-off-by: John Pittman <jpittman@redhat.com>
Signed-off-by: Bryn M. Reeves <bmr@redhat.com>