[openstack_*] work around containerised deployment packaging
Since OpenStack components can now be containerised, and since the
corresponding packages will be present in the container package
manager, rather than the host, OpenStack plugins will fail to run
on these environments since the package triggers do not appear to
be present.

Work around this temporarily by triggering all these plugins on
the presence of the 'openstack-selinux' package: this will always
be installed on the host, regardless of container use, and will
result in the plugins running correctly.

In the future a mechanism to trigger plugins either by packages
installed in a container, or by the presence of specific named
containers will allow this functionality to be restored (and will
prevent all OpenStack plugins from attempting to run even though
some components are absent).

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>