[openstack_*] remove bogus in-plugin rpm --verify
The main Plugin class has supported verify_packages lists for
years: remove the bogus incocations of rpm -V from these plugins
(esp. since they are in the base plugin class, and not specific
to Red Hat or other RPM-using distros), and replace them with a
verify_packages list.

For plugins that may be used with containerised deployments the
verify_packages lists is set from the original list of packages
specific to the component (rather than the 'openstack-selinux'
package that is used for triggering). This enables full verify
support for non-containerised deployments.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>