[openstack_*] restore package verification for OpenStack plugins
Now that the containerised trigger changes have been implemented,
and the in-plugin package verification removed, restore package
verification for these plugins by setting the verify_packages
member of the plugin. The list used is the original package list
for non-containerised deployments (meaning that these types of
system will continue to verify packages as normal, although we
still cannot verify the status of packages inside a container
deployment).

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>