[ovirt] Collect engine heap dump only if it's specified
Currently the java heap dump of engine process is collected from
"/var/log/ovirt-engine/dump" by default. The size of these dumps are
very large which will increase the size of collected sosreport
significantly. The dumps are not required for most of the cases
and hence collect it only if it's specified using -k ovirt.heapdump

Resolves: #1415

Signed-off-by: Nijin Ashok <nashok@redhat.com>
Signed-off-by: Bryn M. Reeves <bmr@redhat.com>