[logging] Always cleanup temp directory
When running `sosreport -l`, or the more specific --list-* options,
sos still creates a temp directory and log files, however we were
not cleaning those up - so over time we could have a multitude of
sos temp directories littering the filesystem.

Now always cleanup the temp directory after execution. Only
exception to this would be if we hit an OSError during execution
_and_ the debug option is specified.

Resolves: #1456

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>
Signed-off-by: Bryn M. Reeves <bmr@redhat.com>