[utilities] Remove the __path__ attribute comparison with empty string
This patch removes an unnecessary comparison on package's __path__
attribute with an empty string in get_modules method.

When we pass an empty path to _find_plugins_from_list function it
returns None. The NoneType is not an iterable entity, so this
will leads to runtime error when we extend the plugins list with
None value.
Signed-off-by: Sourabh Jain <sourabhjain@linux.ibm.com>