[podman|docker] Add postprocessing for container inspect output
Adds a postproc for the podman and docker plugins to attempt to
obfuscate sensitive keys in 'inspect' output for those runtimes.
Previously, these keys were being captured in plaintext which could lead
to passwords or similar being leaked when sysadmins configure containers
with environment variables (or similar) that contain this data.

Specifically, we match against 'key=value' pairs as that is how the
container runtimes accept and print these pairs, like so:

            "Env": [
                "mypassword=supersecret",
                "container=oci"
            ],

By comparison, the inspect outputs now read like the following when a
potentially sensitive key is found:

            "Env": [
                "mypassword=********",
                "container=oci"
            ],

Resolves: #1487

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>
Signed-off-by: Bryn M. Reeves <bmr@redhat.com>