[sosreport] Clear nested threadpool threads on timeout
We nest ThreadPoolExecutors to run plugins to enable whole-plugin
timeouts, with the timeout being handled by the nested threadpool.
However, calling shutdown() on the nested thread is incorrect in the
event of a plugin timeout, as it only cancels pending jobs, not
currently running jobs that the timed-out plugin is.

Instead, call _threads.clear() to forcibly cancel the running job, thus
allowing the thread the return to the higher level threadpool and allow
other plugins to run via that thread.

This still allows data captured up to the point of the timeout to be
written to the sosreport.

Fixes: #1446
Resolves: #1492

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>
Signed-off-by: Bryn M. Reeves <bmr@redhat.com>