[sosreport] Allow users to set timeout for plugins
Previously, plugin timeouts were only able to be set in the plugin
itself, and could not be controlled by users. Now, a new timeout option
has been added to allow this. If a user specifies a timeout, then that
timeout applies to all plugins being run. Setting a timeout value of 0
(zero) means that no timeout will be applied.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>