[kubernetes] Fix 'describe' option enablement and function
The 'describe' option logic was incorrectly nested behind the 'all'
option logic - so users couldn't successfully use the 'describe' option
unless they also specified 'all'.

Fixes the alignment so that 'describe' can be used properly and
separately.

Additionally fixes an issue where the syntax for getting resources for a
specified namespace only worked for pods, and otherwise generated an
error.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>