[Plugin] fix get_option() on Python3 runtimes
The get_option() method uses an internal helper function, _check(),
that tests for matching options that are either strings (equality
test), or iterable types (value present test).

The function assumes that if an __iter__ attribute is present that
it is dealing with an iterable - a list or tuple for e.g.. In the
Python3 string classes this attribute is present (since a string
can be iterated to obtain the characters it contains).

Fix the test by instead testing for the presence of the 'encode'
method which is present on all string types in supported Python 2
and 3 runtimes.

Resolves: #1498.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>