[Plugin] clean up Plugin.get_option()
There's a lot of ancient junk in this method (and associated code
strewn around sos.sosreport and tests). Remove the ability to pass
a list of options to the method since nothing uses this, and also
delete the incomplete implementation of global plugin options via
the commons dictionary (this work was already completed some time
ago by mapping these options directly to the command line args).

Resolves: #1498

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>