[redhat] Add a 'cantboot' preset
Adds a 'cantboot' preset to enable a small number of plugins that may be
useful when troubleshooting a situation in which normal system startup
is failing.

Longer term, it would be beneficial if this preset could be
automatically loaded if sos detects that it is running in a system that
has been booted to rescue/emergency mode.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>