[python] Update plugin to run properly on RHEL 8
RHEL 8 replaces the standard 'python' binary with platform-python,
following PEP394. RHEL 8 sysadmins can however install python2 and/or
python3 alongside this system installation of python and optionally also
set 'python' to point to either of these installations.

This updates the python plugin to enable on the presence of any of these
packages and also get version information for any version of python
installed, including the platform-python installation.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>