[policy] Add a 'none' preset to Policy to override automatic presets
Adds a 'none' preset to the base Policy class that can be used to
override any distro-specific presets that may be otherwise automatically
loaded when a bare 'sosreport' is run due to the presence of certian
packages. In other words, using '--preset=none' will cause sosreport to
run identically to how sosreport ran before presets were introduced.

Additionally this 'none' preset is now the default preset returned by
the Policy class if probe_preset() is not overridden by individual
policies.

Resolves: #1537

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>
Signed-off-by: Bryn M. Reeves <bmr@redhat.com>