[lvm2] Use metadata_read_only=1 for lvm commands
We use locking_type=0 to ensure that we can collect data even if another
process holds a conflicting lock. This introduce a small risk since even
the read-only commands we use may attempt to recover on-disk data in
some cases. The risk is bigger when running lvm2 plugin in oVirt system,
when LVM is used on shared storage, accessible from many hosts.

Add metadata_read_only=1 to ensure that no command can attempt to do
unwanted on-disk changes.

Fixes: #1533
Resolves: #1538

Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Signed-off-by: Bryn M. Reeves <bmr@redhat.com>