[kdump] Collect vmcore-dmesg.txt files based on path and device
kdump allows to configure to which filesystem and what path it will
write contents of '/proc/vmcore'. kdump plugin currently only checks against
a hardcoded path '/var/crash/*/vmcore-dmesg.txt'.

With this change it will check the 'path' variable and also
filesystems type ext2, ext3, ext4 or xfs against fstab. This will make
the plugin check in the correct path in the system.

Signed-off-by: Mikel Olasagasti Uranga <mikel@olasagasti.info>