[postgresql] Use postgres 10 scl if installed
Updates the plugin to check if the specified SCL is running, as some
systems may have multiple SCL versions installed, but only one will be
running at a time. We now use the running version for a pgdump.

This is primarily aimed at RHV environments as 4.3 and later use version
10.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>