[Plugin] allow setting default plugin predicates
Allow plugin authors to set default predicates to apply to either
all collection, or to all command collection, subsequent to the
call to Plugin.set_predicate().

The set predicate will be tested in any subsequent collection call
(depending on the type of predicate), and will gate the collection
of data requested by those calls.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>