[mysql] Collect galera cluster node state
This patch collects the galera mysql node state
stored in the following file:

/var/lib/mysql/grastate.dat

It contains details of the galera node like version,
uuid, seqno, and current value in that node of
safe_to_bootstrap. The idea is that for a galera
mysql cluster, we'll get this status file from
each node in the sosreport, and we'll have a fuller
picture of the issue.

Resolved: #1577

Signed-off-by: Jose Castillo <jcastillo@redhat.com>
Signed-off-by: Bryn M. Reeves <bmr@redhat.com>