[podman] Correct command collection, add volume info collection
Updates the podman plugin to collect volume information such as 'volume
ls' and 'volume inspect' for every reported volume.

Additionally, corrects the inspect iteration to not potentially call
'podman logs' on image IDs, only container IDs.

The image inspect commands will also now use the image name unless
either the name or tage includes 'none', in which case the image ID is
used.

Removes the attempt to collect a podman journal - podman does not log to
a journal unit.

Finally adds '--all' to the stats command, as podman requires this
option when not specifying specific containers.

Resolves: #1579

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>
Signed-off-by: Bryn M. Reeves <bmr@redhat.com>